+++ pybamm/simulation.py
@@ -81,12 +81,6 @@ class Simulation:
         self._parameter_values = parameter_values or model.default_parameter_values
         self._unprocessed_parameter_values = self._parameter_values
 
-        if isinstance(model, pybamm.lithium_ion.BasicDFNHalfCell):
-            if experiment is not None:
-                raise NotImplementedError(
-                    "BasicDFNHalfCell is not compatible with experiment simulations."
-                )
-
         if experiment is None:
             # Check to see if the current is provided as data (i.e. drive cycle)
             current = self._parameter_values.get("Current function [A]")
@@ -558,9 +552,6 @@ class Simulation:
             all_first_states = starting_solution_first_states
             current_solution = starting_solution or pybamm.EmptySolution()
 
-            voltage_stop = self.experiment.termination.get("voltage")
-            logs["stopping conditions"] = {"voltage": voltage_stop}
-
             idx = 0
             num_cycles = len(self.experiment.cycle_lengths)
             feasible = True  # simulation will stop if experiment is infeasible
@@ -829,7 +820,7 @@ class Simulation:
                             capacity_stop = value / 100 * capacity_start
                     else:
                         capacity_stop = None
-                    logs["stopping conditions"]["capacity"] = capacity_stop
+                    logs["stopping conditions"] = {"capacity": capacity_stop}
 
                 logs["elapsed time"] = timer.time()
                 callbacks.on_cycle_end(logs)
@@ -839,11 +830,6 @@ class Simulation:
                 if capacity_stop is not None:
                     capacity_now = cycle_sum_vars["Capacity [A.h]"]
                     if not np.isnan(capacity_now) and capacity_now <= capacity_stop:
-                        break
-
-                if voltage_stop is not None:
-                    min_voltage = cycle_sum_vars["Minimum voltage [V]"]
-                    if min_voltage <= voltage_stop[0]:
                         break
 
                 # Break if the experiment is infeasible (or errored)